feat(gnovm): split mempackage storage into prod and test blobs (#5891)
## What this does

When you deploy a Gno package, the chain stores all of its `.gno` files
together in one blob — the real code *and* the test files (`_test.gno`,
`_filetest.gno`).

The problem: when another package imports yours, the chain has to read
and
type-check the imported package's stored files — including its test
files,
which an importer never actually uses. So importing a package that ships
large
test files costs more gas than it should, and that cost is paid again by
every
importer.

This PR stores each package as **two** blobs instead of one:

- **production files** stay at the normal key `pkg:<path>`
- **test/filetest files** move to a sibling key `pkg:<p

Importing a package now reads only the production blob,
files no longer inflate anyone's import gas. Tools that need the whole
package
(explorers, `gno` tooling) still get everything — the t
together on a full read.

## Also included

A few small, related correctness fixes that come with the split:

- **Reject packages that have no production `.gno` files** (e.g. a
package
  containing only `_test.gno` files). With the split th
blob, which would make them behave differently after a node restart.
They
  have no on-chain use, so they're now rejected at depl
- **Clear old blobs when a private package is redeployed**, so stale
test files
  from a previous version can't linger.
- **Guard a package-path lookup** against odd query input containing
`#`.

## Heads up: this changes the app hash

Because stored bytes change, the genesis app hash changes. Behavior is
unchanged — only the storage encoding shifted. Shipping
relaunch (genesis replay), the same as any storage-format change. The
pinned
app-hash and gas test values are updated accordingly.

## Testing

- New `store_test.go` cases: the two-blob split round-t
redeploy clears stale blobs, and path listing de-duplicates the sibling
key.
- New `addpkg_import_testdep_gas.txtar`: proves importi
test files costs the same gas as importing one **without** (the whole
point of
  the split), plus a matching Go-level equality test.
- All standard suites pass: `gno.land/pkg/sdk/vm`, the txtar integration
suite,
  and the gnovm filetests.

## Files changed

- **Storage (gnovm):** `store.go` (the split), `store_test.go`,
`machine.go`
  (skip rebuilding nodes for production-less packages), `debugger.go`
  (read the sibling key; fixes a possible nil dereference)
- **Keeper (gno.land):** `keeper.go` (clear blobs on private redeploy;
reject
  production-less packages), `keeper_test.go`
- **Fixtures:** new `addpkg_import_testdep_gas.txtar`; re-pinned
  `gnokey_gasfee.txtar`, `restart_gas.txtar`, and the app-hash test

---------

+++ gno.land/pkg/sdk/vm/params.go
@@ -41,7 +41,7 @@ const (
 	minSetReadDepth100Default = int64(200) // 2.0 SET read ops (descent, measured with 10K cache)
 	minWriteDepth100Default   = int64(540) // 4.4 batched COW writes + 1.0 fast-index write
 	// Iterator step flat cost; mirrors store.DefaultGasConfig().IterNextCostFlat.
-	iterNextCostFlatDefault = int64(1_000)
+	iterNextCostFlatDefault = 1_000
 )
 
 var ASCIIDomain = regexp.MustCompile(`^(?:[A-Za-z0-9](?:[A-Za-z0-9-]{0,61}[A-Za-z0-9])?\.)+[A-Za-z]{2,}$`)
@@ -166,7 +166,7 @@ func (p Params) Validate() error {
 	// realistic per-step cost while far below the block gas limit
 	// (~3B) so that a single adversarial proposal can't make one
 	// iterator step drain an entire block's gas budget.
-	const maxIterNextCostFlat = int64(100_000)
+	const maxIterNextCostFlat = 100_000
 	if p.IterNextCostFlat <= 0 {
 		return fmt.Errorf("IterNextCostFlat must be positive, got %d", p.IterNextCostFlat)
 	}
